🏗️🚚`ActiveUserService` moved to the new module.
