+++ business/src/main/java/hu/bmiklos/bc/business/repository/UserRepository.java
@@ -2,7 +2,11 @@ package hu.bmiklos.bc.business.repository;
 
 import hu.bmiklos.bc.domain.entities.User;
 import java.util.Collection;
+import java.util.Optional;
+import java.util.UUID;
 
 public interface UserRepository {
   Collection<User> findAll();
+
+  Optional<User> findById(UUID id);
 }
