+++ business/src/main/java/hu/bmiklos/bc/business/security/NotAuthenticatedException.java
@@ -1,0 +1,3 @@
+package hu.bmiklos.bc.business.security;
+
+public class NotAuthenticatedException extends RuntimeException {}
