--- src/main/java/hu/bmiklos/bc/exception/NotAuthenticatedException.java
@@ -1,5 +1,0 @@
-package hu.bmiklos.bc.exception;
-
-public class NotAuthenticatedException extends RuntimeException {
-    
-}
